fix warnings

+++ Source/VTKExtensions/Filtering/vtkCMBArcCreateOperator.cxx
@@ -73,10 +73,10 @@ bool vtkCMBArcCreateOperator::Operate(vtkPolyData *source)
     for ( vtkIdType i=0; i < ids->GetNumberOfIds(); ++i)
       {
       source->GetPoint(ids->GetId(i),pos);
-      unsigned int pointID = i;
+      unsigned int pointID = static_cast<unsigned int>(i);
       if(vda != NULL)
         {
-        pointID = vda->GetTuple1(ids->GetId(i));
+        pointID = static_cast<unsigned int>(vda->GetTuple1(ids->GetId(i)));
         }
       vtkCMBArc::Point tmpPt(pos, pointID);
       if (currentIndex != 0 && currentIndex != numberOfInternalEndNodes-1)
